Move the invokers to separate files to facilitate their reuse
